moved calc elapsed avail from User to UserCtrl
